--- packages/app/src/library/PluginLabel/types.ts
@@ -1,8 +1,0 @@
-// Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import type { Plugin } from 'config/plugins'
-
-export interface PluginLabelProps {
-  plugin: Plugin
-}
