+++ packages/app/src/library/ValidatorList/Item.tsx
@@ -1,12 +1,10 @@
 // Copyright 2025 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import { faBars, faGlobe, faTimes } from '@fortawesome/free-solid-svg-icons'
+import { faTimes } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import type { AnyJson } from '@w3ux/types'
 import classNames from 'classnames'
 import { useList } from 'contexts/List'
-import { useMenu } from 'contexts/Menu'
 import { usePlugins } from 'contexts/Plugins'
 import { CurrentEraPoints } from 'library/List/EraPointsGraph/CurrentEraPoints'
 import { HistoricalEraPoints } from 'library/List/EraPointsGraph/HistoricalEraPoints'
@@ -16,15 +14,12 @@ import { Metrics } from 'library/ListItem/Labels/Metrics'
 import { ParaValidator } from 'library/ListItem/Labels/ParaValidator'
 import { Quartile } from 'library/ListItem/Labels/Quartile'
 import { Wrapper } from 'library/ListItem/Wrappers'
-import { MenuList } from 'library/Menu/List'
-import { useTranslation } from 'react-i18next'
 import {
   HeaderButton,
   HeaderButtonRow,
   LabelRow,
   Separator,
 } from 'ui-core/list'
-import { useOverlay } from 'ui-overlay'
 import { useValidators } from '../../contexts/Validators/ValidatorEntries'
 import { Blocked } from '../ListItem/Labels/Blocked'
 import { Commission } from '../ListItem/Labels/Commission'
@@ -41,46 +36,11 @@ export const Item = ({
   eraPoints,
   removeHandler,
 }: ItemProps) => {
-  const { t } = useTranslation('app')
-  const { openMenu, open } = useMenu()
   const { pluginEnabled } = usePlugins()
-  const { openModal } = useOverlay().modal
   const { selectable, selected } = useList()
   const { validatorIdentities, validatorSupers } = useValidators()
   const { address, prefs, validatorStatus } = validator
   const commission = prefs?.commission ?? null
-
-  const identity = getIdentityDisplay(
-    validatorIdentities[address],
-    validatorSupers[address]
-  ).node
-
-  // Configure menu
-  const menuItems: AnyJson[] = []
-
-  if (pluginEnabled('polkawatch')) {
-    menuItems.push({
-      icon: <FontAwesomeIcon icon={faGlobe} transform="shrink-3" />,
-      wrap: null,
-      title: `${t('viewDecentralization')}`,
-      cb: () => {
-        openModal({
-          key: 'ValidatorGeo',
-          options: {
-            address,
-            identity,
-          },
-        })
-      },
-    })
-  }
-
-  // Handler for opening menu
-  const toggleMenu = (ev: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    if (!open) {
-      openMenu(ev, <MenuList items={menuItems} />)
-    }
-  }
 
   const isSelected = !!selected.filter(
     (item) => item.address === validator.address
@@ -101,13 +61,6 @@ export const Item = ({
             <HeaderButtonRow>
               <CopyAddress address={address} />
               {toggleFavorites && <FavoriteValidator address={address} />}
-              {!['modal', 'canvas'].includes(displayFor) && (
-                <HeaderButton>
-                  <button type="button" onClick={(ev) => toggleMenu(ev)}>
-                    <FontAwesomeIcon icon={faBars} transform="shrink-2" />
-                  </button>
-                </HeaderButton>
-              )}
               {displayFor === 'default' && (
                 <Metrics
                   address={address}
