+++ packages/locales/src/resources/en/app.json
@@ -279,7 +279,6 @@
     "validatorPerformance": "{{count}} Day Validator Performance",
     "validators": "Validators",
     "valueTooSmall": "Value is too small",
-    "viewDecentralization": "View Decentralization",
     "viewPoolNominations": "View Pool Nominations",
     "waiting": "Waiting",
     "walletNotFound": "wallet not found",
