+++ packages/locales/src/resources/en/modals.json
@@ -321,7 +321,6 @@
     "updatePoolCommission": "Update pool commission settings.",
     "updateWhoClaimRewards": "Update who can claim rewards on your behalf.",
     "updated": "Updated",
-    "validatorDecentralization": "Validator Decentralization",
     "validatorMetrics": "Validator Metrics",
     "viewPoolSubtitle": "View pool performance and other metrics.",
     "viewPoolTitle": "View Pool Metrics",
