+++ packages/locales/src/resources/es/app.json
@@ -279,7 +279,6 @@
     "validatorPerformance": "{{count}} Días de Rendimiento del Validador",
     "validators": "Validadores",
     "valueTooSmall": "El valor es demasiado pequeño",
-    "viewDecentralization": "Ver Descentralización",
     "viewPoolNominations": "Ver Nominaciones del Pool",
     "waiting": "Esperando",
     "walletNotFound": "Cartera no encontrada",
