+++ packages/locales/src/resources/es/modals.json
@@ -321,7 +321,6 @@
     "updatePoolCommission": "Actualizar configuración de comisión del Pool.",
     "updateWhoClaimRewards": "Actualizar quién puede reclamar recompensas en tu nombre.",
     "updated": "Actualizado",
-    "validatorDecentralization": "Descentralización de Validadores",
     "validatorMetrics": "Métricas de Validadores",
     "viewPoolSubtitle": "Vea el rendimiento del grupo y otras métricas.",
     "viewPoolTitle": "Ver métricas del pool",
