+++ packages/locales/src/resources/zh/app.json
@@ -270,7 +270,6 @@
     "validatorPerformance": "{{count}}天内验证人表现",
     "validators": "验证人",
     "valueTooSmall": "值太小",
-    "viewDecentralization": "分布式指标",
     "viewPoolNominations": "查看池提名",
     "waiting": "等待中",
     "walletNotFound": "未找到钱包",
