+++ packages/locales/src/resources/zh/modals.json
@@ -307,7 +307,6 @@
     "updatePoolCommission": "更新提名池佣金设置",
     "updateWhoClaimRewards": "更改代表申领奖励地址",
     "updated": "己更新",
-    "validatorDecentralization": "验证人的分布式地理指标",
     "validatorMetrics": "验证人指标",
     "viewPoolSubtitle": "查看池性能和其他指标.",
     "viewPoolTitle": "查看池指标",
